- implemented changing baackground color through all the app - added Provider package - minor refactoring
