+++ relax_button/lib/screens/home/home_screen.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
 import 'package:relax_button/constants/routes.dart';
 import 'package:relax_button/constants/texts.dart';
+import 'package:relax_button/models/background_color_model.dart';
 
 class HomeScreen extends StatelessWidget {
   const HomeScreen({Key? key}) : super(key: key);
@@ -9,6 +11,8 @@ class HomeScreen extends StatelessWidget {
   Widget build(BuildContext context) {
     // TODO: center widget is a temp
     return Container(
+      // TODO: save this color somewhere./refactor
+      color: Provider.of<BackgroundColorModel>(context).color,
       child: Column(
         children: [
           ElevatedButton(
