+++ relax_button/pubspec.yaml
@@ -20,6 +20,7 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
+  provider: ^6.0.2
   rive: ^0.8.4
   just_audio: ^0.9.20
